Added main menu and implementation for adding/deleting folders
I added the functionality of the main menu in main and have added the ability to create and delete folders using the folderManager class, I have also implemented the printFolders function as it was helpful in debugging my code. Also implemented the folderIsEmpty function to check if the folders vector is empty. Also set up a lot of the .cpp files with the function names, but have not implemented most of them, many have dummy return values so I wouldn't get errors from the complier.