Added user input authentication for folder menu and username input
Added authentication for user input when the user is asked to input their username, so they are not allowed to have a blank username. Also added implemenetation that checks to see that a new folder cannot be purely whitespace. User input on folder menu is also authenticated so user cannot enter anything other than the menu options. Any other input that is not allowed will print an error and return them to the menu. Also added a new getFolderSize() function to folderManager class so the size of folder is accessible.